Updated PR according to code review - The copy and alert presenting methods are now under UIAlertController extension. - Fixed typo - removed @objcfrom UIViewController extension
