+++ WordPress/Classes/ViewRelated/Notifications/Views/LongPressGestureLabel.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-/// Sublclass of UILabel that includes a long press gesture recognizer.
+/// Subclass of UILabel that includes a long press gesture recognizer.
 /// The action attached to the recognizer will be set by the creator of the label.
 
 class LongPressGestureLabel: UILabel {
